+++ Blog.Core/Blog.Core.xml
@@ -300,6 +300,13 @@
             <param name="BlogArticle"></param>
             <returns></returns>
         </member>
+        <member name="M:Blog.Core.Controllers.BlogController.ApacheTestUpdate">
+            <summary>
+            apache jemeter 压力测试
+            更新接口
+            </summary>
+            <returns></returns>
+        </member>
         <member name="T:Blog.Core.Controllers.ClaimsController">
             <summary>
             这是一个测试的控制器，主要为了测试基于Claim的验证机制
@@ -921,6 +928,23 @@
             全局路由前缀公约
             </summary>
         </member>
+        <member name="T:Blog.Core.Middlewares.IPLogMildd">
+            <summary>
+            中间件
+            记录IP请求数据
+            </summary>
+        </member>
+        <member name="F:Blog.Core.Middlewares.IPLogMildd._next">
+            <summary>
+            
+            </summary>
+        </member>
+        <member name="M:Blog.Core.Middlewares.IPLogMildd.#ctor(Microsoft.AspNetCore.Http.RequestDelegate)">
+            <summary>
+            
+            </summary>
+            <param name="next"></param>
+        </member>
         <member name="M:Blog.Core.Middlewares.MiddlewareHelpers.UseJwtTokenAuth(Microsoft.AspNetCore.Builder.IApplicationBuilder)">
             <summary>
             自定义授权中间件
@@ -945,6 +969,13 @@
         <member name="M:Blog.Core.Middlewares.MiddlewareHelpers.UseExceptionHandlerMidd(Microsoft.AspNetCore.Builder.IApplicationBuilder)">
             <summary>
             异常处理中间件
+            </summary>
+            <param name="app"></param>
+            <returns></returns>
+        </member>
+        <member name="M:Blog.Core.Middlewares.MiddlewareHelpers.UseIPLogMildd(Microsoft.AspNetCore.Builder.IApplicationBuilder)">
+            <summary>
+            IP请求中间件
             </summary>
             <param name="app"></param>
             <returns></returns>
