+++ Blog.Core/appsettings.json
@@ -22,7 +22,7 @@
       "ConnectionString": "127.0.0.1:6319"
     },
     "MemoryCachingAOP": {
-      "Enabled": true
+      "Enabled": false
     },
     "LogAOP": {
       "Enabled": false
@@ -64,19 +64,22 @@
   },
   "Startup": {
     "Cors": {
-      "IPs": "http://127.0.0.1:1818,http://localhost:8080,http://localhost:8021,http://localhost:8081,http://localhost:1818"
+      "IPs": "http://127.0.0.1:2364,http://localhost:2364,http://localhost:8080,http://localhost:8021,http://localhost:1818"
     },
     "ApiName": "Blog.Core"
   },
   "Middleware": {
     "RequestResponseLog": {
+      "Enabled": false
+    },
+    "IPLog": {
       "Enabled": true
     },
     "RecordAllLogs": {
       "Enabled": false
     },
     "SignalR": {
-      "Enabled": true
+      "Enabled": false
     }
   }
 }
